feat(behaviors): caps-word supports alternative layouts
The problem when using alternative layouts (Colemak, Dvorak, etc) mapped
in the OS, or when using a foreign keyboard, there will be alpha
characters that the keyboard "thinks" are symbols, and therefore
caps-word will not shift them. This adds the ability to configure
caps-word to shift these keycodes so that it works as expected.